% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_return.R
\docType{data}
\name{index_return}
\alias{index_return}
\title{Daily log-return of the index values from four countries (United States, Japan, Germany, and South Korea) for the period spanning from May 2009 to September 2009}
\format{
A data frame with 419 rows and 4 variables:
}
\source{
\url{United States (S&P500), Japan (NIKKEI225) , Germany (DAX Index), and South Korea (KOSPI)}
}
\usage{
index_return
}
\description{
A dataframe contains daily log-return of the index values from four countries (United States, Japan, Germany, and South Korea) for the period spanning from May 2009 to September 2009.
}
\keyword{datasets}
