% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_return_four_countries.R
\docType{data}
\name{time_series_data}
\alias{time_series_data}
\title{COVID-19 infection trajectories of 30 countries}
\format{
A data frame with 30 rows and 116 variables:
}
\source{
\url{https://github.com/CSSEGISandData/COVID-19}
}
\usage{
time_series_data
}
\description{
A dataframe contains COVID-19 infection trajectories of 30 countries from Jan 22nd, 2020 to May 14th, 2020.
}
\keyword{datasets}
